+++ src/main/java/kitchenpos/order/application/OrderResponse.java
@@ -1,11 +1,11 @@
-package kitchenpos.application.dto;
+package kitchenpos.order.application;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
 
 public class OrderResponse {
