+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -9,16 +9,16 @@ import static kitchenpos.Fixture.테이블;
 import static kitchenpos.Fixture.테이블집합;
 
 import java.math.BigDecimal;
-import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.TableGroupDao;
-import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuDao;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menugroup.domain.MenuGroupDao;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderDao;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductDao;
 import org.springframework.beans.factory.annotation.Autowired;
